# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code from the catalog.",
            "title": "Code",
            "type": "string"
          },
          "invalid_params": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field detail, present only on validation failures.",
            "title": "Invalid Params"
          },
          "key": {
            "description": "Localization key; clients may resolve their own text from it.",
            "title": "Key",
            "type": "string"
          },
          "message": {
            "description": "English message. Localized clients should prefer key + params.",
            "title": "Message",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "description": "Values interpolated into the localized message.",
            "title": "Params",
            "type": "object"
          },
          "severity": {
            "description": "Catalog severity: INFO, WARNING, ERROR or CRITICAL.",
            "title": "Severity",
            "type": "string"
          },
          "trace_id": {
            "description": "Correlates the failure with the audit trail.",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "severity",
          "key",
          "params",
          "message",
          "trace_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ProxyProviderConfigResponse": {
        "description": "The tenant's proxy provider configuration.\n\nReturned identically by the read and the upsert, so one model serves both.\n`api_key_masked` follows the same rule as the LLM family: a mask or null,\nnever the stored key, which is encrypted at rest.",
        "properties": {
          "api_key_masked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Masked provider key, or null when the provider needs none.",
            "title": "Api Key Masked"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO-8601 UTC with a Z suffix.",
            "title": "Created At"
          },
          "default_model": {
            "title": "Default Model",
            "type": "string"
          },
          "provider": {
            "description": "`openai`, `ollama` or `custom`.",
            "title": "Provider",
            "type": "string"
          },
          "timeout_seconds": {
            "title": "Timeout Seconds",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "provider",
          "base_url",
          "api_key_masked",
          "default_model",
          "timeout_seconds",
          "created_at",
          "updated_at"
        ],
        "title": "ProxyProviderConfigResponse",
        "type": "object"
      },
      "ProxySettingsPutSchema": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "default_model": {
            "title": "Default Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "timeout": {
            "default": 60,
            "title": "Timeout",
            "type": "integer"
          }
        },
        "required": [
          "provider",
          "base_url",
          "default_model"
        ],
        "title": "ProxySettingsPutSchema",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI Security Gateway for Intelligent Applications",
    "title": "WrapSec",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/settings/proxy": {
      "delete": {
        "description": "Removes the proxy provider config for the current tenant.\nAfter deletion, proxy mode requests will fail with proxy_not_configured.\n404 if no config exists.",
        "operationId": "delete_proxy_settings_v1_settings_proxy_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Proxy provider configuration is admin-only, on reads as well as writes."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "This tenant has no proxy provider configured."
          }
        },
        "summary": "Delete Proxy Settings",
        "tags": [
          "Proxy"
        ]
      },
      "get": {
        "description": "Returns the proxy provider config for the current tenant.\nThe provider API key is always masked in the response - never returned in plaintext.\n404 if no provider has been configured.",
        "operationId": "get_proxy_settings_v1_settings_proxy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyProviderConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Proxy provider configuration is admin-only, on reads as well as writes."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "This tenant has no proxy provider configured."
          }
        },
        "summary": "Get Proxy Settings",
        "tags": [
          "Proxy"
        ]
      },
      "put": {
        "description": "Creates or replaces the proxy provider config for the current tenant (upsert).\nThe provider API key is encrypted before storage using the server's secret_key.\nProviders \"openai\" and \"custom\" require api_key; \"ollama\" does not.",
        "operationId": "put_proxy_settings_v1_settings_proxy_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxySettingsPutSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyProviderConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Proxy provider configuration is admin-only, on reads as well as writes."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request body failed validation."
          }
        },
        "summary": "Put Proxy Settings",
        "tags": [
          "Proxy"
        ]
      }
    },
...
  }
}